Reintroduce OLD_ARRAY of UINT32 -> OLD_ARRAY of OBJECT_ID mapping (#38)
Before DevTools was based on UCSL and rip, we had an optimization in
the reflection editor specifically for Sonic Forces, where any reflections
for OLD_ARRAY arrays of UINT32 items were automatically converted to
OLD_ARRAY arrays of OBJECT_ID items.

This optimization was added because until now we have only observed OLD_ARRAY
in the context of Forces' gedit v2, where OLD_ARRAYs of UINT32s are only used
for lists of object IDs. The optimization (or hack if your prefer) made it so
that these lists would be properly handled as lists of object IDs instead of
random integers.

During the migration to UCSL/rip this optimization was lost, breaking the
Sonic Forces version of DevTools. This commit reintroduces the optimization.